CLC-5837 OEC: Reorganize validate/export/publish tasks
